move report routes to blueprints/reports.py
