+++ app/templates/favourites.html
@@ -44,7 +44,7 @@
                         {% endif %}
 
                         <div class="fav-panel-actions">
-                            <a class="btn btn-ghost" href="{{ url_for('listing_page') }}?city_id={{ s.id }}">View Reports</a>
+                            <a class="btn btn-ghost" href="{{ url_for('reports.listing_page') }}?city_id={{ s.id }}">View Reports</a>
                         </div>
                     </div>
                 </article>
