+++ app/templates/profile.html
@@ -168,7 +168,7 @@
             {% for r in recent_reports %}
                 <li class="profile-post-card" style="--cat: {{ r.category.marker_color }};">
                     {# whole-card stretched link — Edit + cat-tag escape via z-index #}
-                    <a class="profile-post-link" href="{{ url_for('view_report', token=r.id|report_token) }}" aria-label="View report"></a>
+                    <a class="profile-post-link" href="{{ url_for('reports.view_report', token=r.id|report_token) }}" aria-label="View report"></a>
 
                     <div class="profile-post-header">
                         <span class="profile-post-icon" aria-hidden="true">{{ CATEGORY_EMOJI.get(r.category.name, '📝') }}</span>
@@ -191,8 +191,8 @@
                         {% set card_is_own = current_user.is_authenticated and r.user_id == current_user.id %}
                         {% if card_is_own %}
                             <div class="own-report-actions">
-                                <a href="{{ url_for('view_report', token=r.id|report_token) }}" class="own-report-tag">Your report</a>
-                                <a href="{{ url_for('edit_report_page', report_id=r.id) }}" class="listing-item-edit">
+                                <a href="{{ url_for('reports.view_report', token=r.id|report_token) }}" class="own-report-tag">Your report</a>
+                                <a href="{{ url_for('reports.edit_report_page', report_id=r.id) }}" class="listing-item-edit">
                                     <svg xmlns="http://www.w3.org/2000/svg" width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
                                         <path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"/>
                                         <path d="M18.5 2.5a2.121 2.121 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5z"/>
@@ -286,7 +286,7 @@
                     {# stats footer — likes/comments are stubs until those features exist #}
                     <div class="profile-post-stats">
                         <a class="profile-post-stat profile-post-stat-link"
-                           href="{{ url_for('view_report', token=r.id|report_token) }}#comments"
+                           href="{{ url_for('reports.view_report', token=r.id|report_token) }}#comments"
                            title="View comments">
                             <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
                                 <path d="M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z"/>
