+++ app/templates/report_view.html
@@ -7,7 +7,7 @@
 
 {% block content %}
 <section class="report-view-page">
-    <a href="{{ url_for('listing_page') }}" class="report-view-back">← Back to all reports</a>
+    <a href="{{ url_for('reports.listing_page') }}" class="report-view-back">← Back to all reports</a>
 
     <article class="report-view-card" style="--cat: {{ report.category.marker_color }};">
         <header class="report-view-header">
@@ -112,7 +112,7 @@
             {% if is_own %}
                 <div class="own-report-actions">
                     <span class="own-report-tag">Your report</span>
-                    <a href="{{ url_for('edit_report_page', report_id=report.id) }}" class="report-view-edit-btn">
+                    <a href="{{ url_for('reports.edit_report_page', report_id=report.id) }}" class="report-view-edit-btn">
                         <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
                             <path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"/>
                             <path d="M18.5 2.5a2.121 2.121 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5z"/>
@@ -322,7 +322,7 @@
         try {
             const res = await fetch(`/api/reports/${btn.dataset.reportId}`, {method: 'DELETE'});
             if (res.ok) {
-                window.location.href = "{{ url_for('listing_page') }}";
+                window.location.href = "{{ url_for('reports.listing_page') }}";
             } else {
                 btn.disabled = false;
                 alert('Could not delete the report. Please try again.');
