+++ app/templates/reports_listing.html
@@ -16,7 +16,7 @@
             {% endif %}
         </h1>
         {% if current_user.is_authenticated %}
-            <a href="{{ url_for('reports_page') }}" class="listing-new-btn">+ New Report</a>
+            <a href="{{ url_for('reports.reports_page') }}" class="listing-new-btn">+ New Report</a>
         {% else %}
             {# guest — point at signup so they can join and post #}
             <a href="{{ url_for('auth.signup') }}" class="listing-new-btn">Sign up to post</a>
@@ -193,7 +193,7 @@
                        Edit button + author profile link sit above it via z-index. #}
                     <div class="listing-item-text">
                         <a class="listing-item-link"
-                           href="{{ url_for('view_report', token=r.id|report_token) }}"
+                           href="{{ url_for('reports.view_report', token=r.id|report_token) }}"
                            aria-label="View report"></a>
  
                         <div class="listing-item-top">
@@ -242,8 +242,8 @@
                                     data-saved="{{ 'true' if is_saved else 'false' }}">
                                 {{ 'Saved ✓' if is_saved else 'Save' }}
                             </button>
-                            <a href="{{ url_for('view_report', token=r.id|report_token) }}" class="own-report-tag">Your report</a>
-                            <a href="{{ url_for('edit_report_page', report_id=r.id) }}" class="listing-item-edit">
+                            <a href="{{ url_for('reports.view_report', token=r.id|report_token) }}" class="own-report-tag">Your report</a>
+                            <a href="{{ url_for('reports.edit_report_page', report_id=r.id) }}" class="listing-item-edit">
                                 <svg xmlns="http://www.w3.org/2000/svg" width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
                                     <path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"/>
                                     <path d="M18.5 2.5a2.121 2.121 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5z"/>
@@ -357,7 +357,7 @@
         {% endif %}
     {% else %}
         <div class="listing-empty">
-            No reports yet. Be the first to <a href="{{ url_for('reports_page') }}">submit one</a>.
+            No reports yet. Be the first to <a href="{{ url_for('reports.reports_page') }}">submit one</a>.
         </div>
     {% endif %}
 </section>
